Refactoring load products - creating use case
