+++ src/domain/Product.ts
@@ -1,0 +1,6 @@
+export interface Product {
+    id: number;
+    title: string;
+    image: string;
+    price: string;
+}
