+++ src/presentation/hooks/useProducts.ts
@@ -1,41 +1,19 @@
 import { useEffect, useState } from "react";
-import { RemoteProduct, StoreApi } from "../../data/api/StoreApi";
 import { useReload } from "./useReload";
+import { GetProcductsUseCase } from "../../domain/GetProductsUseCase";
+import { Product } from "../../domain/Product";
 
-export function useProducts(storeApi: StoreApi) {
+export function useProducts(_getProductsUseCase: GetProcductsUseCase) {
     const [reloadKey, reload] = useReload();
 
     const [products, setProducts] = useState<Product[]>([]);
     useEffect(() => {
-        storeApi.getAll().then(response => {
+        _getProductsUseCase.Execute().then(products => {
             console.debug("Reloading", reloadKey);
-
-            const remoteProducts = response as RemoteProduct[];
-
-            const products = remoteProducts.map(buildProduct);
 
             setProducts(products);
         });
-    }, [reloadKey, storeApi]);
+    }, [reloadKey, _getProductsUseCase]);
 
     return { reload, products };
-}
-
-export function buildProduct(remoteProduct: RemoteProduct): Product {
-    return {
-        id: remoteProduct.id,
-        title: remoteProduct.title,
-        image: remoteProduct.image,
-        price: remoteProduct.price.toLocaleString("en-US", {
-            maximumFractionDigits: 2,
-            minimumFractionDigits: 2,
-        }),
-    };
-}
-
-export interface Product {
-    id: number;
-    title: string;
-    image: string;
-    price: string;
 }
